Add video camera & misc changes
apa.py is copied from Raspio.tv. It is loaded here for completeness of the set of files to support the led assemblies used by this project. There is one line of code added to apa.py that seems to be needed for the SPI interface to operate correctly. It is commented in the code for apa.py.